elemental-utils/requirements.txt
pynetbox
six==1.*
dnspython
paramiko
ansible>=2.9.27
pyvmomi
pysftp
pyyaml
hvac
